requirements.txt
Django==5.2.8
dj-database-url==3.0.1
python-decouple==3.8
psycopg2-binary==2.9.11
requests==2.32.5
gunicorn==23.0.0
whitenoise==6.11.0
google-generativeai==0.8.3
